refactor(PATHS): Remove 'destinations' from the src file paths to match the new published artifact.
